+++ src/Sulu/Bundle/ContactBundle/Tests/Functional/Controller/ContactControllerTest.php
@@ -148,6 +148,7 @@ class ContactControllerTest extends SuluTestCase
         $contact->setFirstName('Max');
         $contact->setLastName('Mustermann');
         $contact->setPosition('CEO');
+        $contact->setBirthday(new \DateTime());
         $contact->setFormOfAddress(1);
         $contact->setSalutation('Sehr geehrter Herr Dr Mustermann');
 
@@ -831,7 +832,10 @@ class ContactControllerTest extends SuluTestCase
         $response = json_decode($client->getResponse()->getContent());
 
         $this->assertEquals(400, $client->getResponse()->getStatusCode());
-        $this->assertEquals('The "Sulu\Bundle\ContactBundle\Entity\Contact"-entity requires a "contact"-argument', $response->message);
+        $this->assertEquals(
+            'The "Sulu\Bundle\ContactBundle\Entity\Contact"-entity requires a "contact"-argument',
+            $response->message
+        );
     }
 
     public function testPostWithEmptyAdditionalData()
@@ -2159,6 +2163,30 @@ class ContactControllerTest extends SuluTestCase
         $this->assertEquals(false, $response->addresses[0]->primaryAddress);
         $this->assertEquals(false, $response->addresses[1]->primaryAddress);
         $this->assertEquals(true, $response->addresses[2]->primaryAddress);
+    }
+
+    public function testPostEmptyBirthday()
+    {
+        $client = $this->createTestClient();
+
+        $client->request('GET', '/api/contacts/' . $this->contact->getId());
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertArrayHasKey('birthday', $response);
+
+        $data = [
+            'firstName' => 'John',
+            'lastName' => 'Doe',
+            'title' => $this->contactTitle->getId(),
+            'birthday' => '',
+        ];
+
+        $client->request('PUT', '/api/contacts/' . $this->contact->getId(), $data);
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertArrayNotHasKey('birthday', $response);
+
+        $client->request('GET', '/api/contacts/' . $this->contact->getId());
+        $response = json_decode($client->getResponse()->getContent(), true);
+        $this->assertArrayNotHasKey('birthday', $response);
     }
 
     public function sortAddressesPrimaryLast()